+++ src/templates/blog/login.html
@@ -11,7 +11,7 @@
 <body class="bg">
     <div class="login">
         <div class="avatar" style="background-image: url('https://obmfhf1m3.qnssl.com/Katarina.png');"></div>
-        <form action="/login" method="POST">
+        <form action="/auth/login" method="POST">
             <input type="text" name="username" placeholder="用户名">
             <input type="password" name="password" placeholder="密码">
             <button type="submit">登录</button>
